Separate traverse into a static method that accepts a path list and request method.
This allows us to test its behabior independently of the Metadata module and the AWS.MetadataService client.
