+++ lib/source/metadata/util.js
@@ -1,6 +1,8 @@
 'use strict';
+const Path = require('path');
 
 const DEFAULT_PARALLEL = 16;
+const METADATA_LIST = /^(\d+)=(.+)$/;
 
 /**
  * Iterate over a list of tasks in parallel, passing one to each call of the
@@ -15,7 +17,7 @@ const DEFAULT_PARALLEL = 16;
  * @param  {Function} done    Callback on error or completion
  * @param  {Object}   s { parallel: Number limit parallel tasks. Default 16 }
  */
-exports.each = function each(list, work, done, s) {
+function each(list, work, done, s) {
   const state = s || {
     parallel: DEFAULT_PARALLEL,
     error: false
@@ -74,4 +76,52 @@ exports.each = function each(list, work, done, s) {
     // Iterate
     each(list, work, done, state);
   });
+}
+exports.each = each;
+
+/**
+ * Fetch all values from the EC2 metadata tree
+ *
+ * @param  {Function} request(path, cb)    Request handler. Called for each Metadata path
+ * @param  {Function} callback(err, paths) Return an error or a hash of paths and their values
+ *
+ * @private
+ */
+exports.traverse = function traverse(paths, request, callback) {
+  const values = {};
+
+  each(paths, (path, next) => {
+    request(Path.join('/', this._version, path), (err, data) => {
+      if (err) {
+        return next(err);
+      }
+
+      // This is a tree! Split new-line delimited strings into an array and add to tail of paths
+      if (path.slice(-1) === '/') {
+        const items = data.trim().split('\n');
+
+        // Is this a list?
+        if (items.reduce((memo, item) => memo && METADATA_LIST.test(item), true)) {
+          const list = values[path.slice(1, -1)] = [];
+
+          items.forEach((item) => {
+            const match = item.match(METADATA_LIST);
+
+            if (match) {
+              list[match[1]] = match[2];
+            }
+          });
+
+          return next();
+        }
+
+        items.forEach((node) => paths.push(Path.join(path, node)));
+        return next();
+      }
+
+      // Remove leading `/`
+      values[path.slice(1)] = data;
+      next();
+    });
+  }, (err) => callback(err, values), null);
 };
